:boom: Update url/view/template structure -- #204
