+++ src/sdg/organisaties/views/__init__.py
@@ -1,3 +1,4 @@
+from .catalogi import *  # noqa
 from .invitation import *  # noqa
 from .overheden import *  # noqa
 from .roles import *  # noqa
