+++ src/sdg/organisaties/views/roles.py
@@ -6,7 +6,7 @@ from sdg.organisaties.views.mixins import DisallowOwnRoleMixin, RoleBaseMixin
 
 
 class RoleListView(RoleBaseMixin, OverheidMixin, ListView):
-    template_name = "organisaties/overheid_role_list.html"
+    template_name = "organisaties/roles/list.html"
     required_roles = ["is_beheerder", "is_redacteur"]
 
     def get_queryset(self):
@@ -22,14 +22,14 @@ class RoleListView(RoleBaseMixin, OverheidMixin, ListView):
 
 class RoleDeleteView(DisallowOwnRoleMixin, RoleBaseMixin, OverheidMixin, DeleteView):
     queryset = Role.objects.all()
-    template_name = "organisaties/overheid_role_delete.html"
+    template_name = "organisaties/roles/delete.html"
     pk_url_kwarg = "role_pk"
     required_roles = ["is_beheerder"]
 
 
 class RoleUpdateView(DisallowOwnRoleMixin, RoleBaseMixin, OverheidMixin, UpdateView):
     queryset = Role.objects.all()
-    template_name = "organisaties/overheid_role_update.html"
+    template_name = "organisaties/roles/update.html"
     pk_url_kwarg = "role_pk"
     required_roles = ["is_beheerder"]
     fields = ["is_beheerder", "is_redacteur"]
