+++ src/sdg/producten/urls.py
@@ -9,14 +9,8 @@ from sdg.producten.views.product import (
 app_name = "producten"
 urlpatterns = [
     url(
-        r"^(?P<pk>\d+)/$",
-        ProductDetailView.as_view(),
-        name="detail",
+        r"r/(?P<product_pk>\d+)/", ProductCreateRedirectView.as_view(), name="redirect"
     ),
-    url(r"^(?P<pk>\d+)/edit/$", ProductUpdateView.as_view(), name="edit"),
-    url(
-        r"^r/(?P<pk>\d+)/(?:(?P<catalog_pk>\d+)/)?$",
-        ProductCreateRedirectView.as_view(),
-        name="redirect",
-    ),
+    url(r"(?P<product_pk>\d+)/", ProductDetailView.as_view(), name="detail"),
+    url(r"(?P<product_pk>\d+)/edit/", ProductUpdateView.as_view(), name="edit"),
 ]
