+++ src/sdg/producten/views/product.py
@@ -42,11 +42,18 @@ class ProductCreateRedirectView(SingleObjectMixin, RedirectView):
             else:
                 raise PermissionDenied()
 
-        kwargs["obj"] = obj
+        kwargs["product"] = obj
         return super().get(request, *args, **kwargs)
 
     def get_redirect_url(self, *args, **kwargs):
-        return reverse("producten:detail", kwargs={"pk": kwargs.get("obj").pk})
+        return reverse(
+            "organisaties:catalogi:producten:detail",
+            kwargs={
+                "pk": kwargs.get("product").catalogus.lokale_overheid.pk,
+                "catalog_pk": kwargs.get("product").catalogus.pk,
+                "product_pk": kwargs.get("product").pk,
+            },
+        )
 
 
 class ProductDetailView(OverheidMixin, DetailView):
@@ -162,4 +169,11 @@ class ProductUpdateView(OverheidMixin, UpdateView):
         )
 
     def get_success_url(self):
-        return reverse("producten:detail", kwargs={"pk": self.product.pk})
+        return reverse(
+            "organisaties:catalogi:producten:detail",
+            kwargs={
+                "pk": self.lokale_overheid.pk,
+                "catalog_pk": self.product.catalogus.pk,
+                "product_pk": self.product.pk,
+            },
+        )
