+++ src/sdg/templates/_include/product_catalog.html
@@ -18,10 +18,11 @@
                 <div class="products__list-container">
                     {% for product in producten %}
                         {% if catalog.is_referentie_catalogus %}
-                            <a class="products__item" href="{% url 'producten:detail' pk=product.pk %}">
+                            <a class="products__item"
+                               href="{% url 'organisaties:catalogi:producten:detail' pk=lokaleoverheid.pk catalog_pk=catalog.pk product_pk=product.pk %}">
                         {% else %}
                             <a class="products__item"
-                               href="{% url 'producten:redirect' pk=product.pk catalog_pk=catalog.pk %}">
+                               href="{% url 'organisaties:catalogi:producten:redirect' pk=lokaleoverheid.pk catalog_pk=catalog.pk product_pk=product.pk %}">
                         {% endif %}
                     <div class="products__item-icon">
                         {% product_icon product %}
