+++ src/sdg/templates/producten/_include/product_status.html
@@ -23,7 +23,8 @@
         {% endwith %}
     </div>
     <div>
-        <a class="button" href="{% url 'producten:edit' pk=product.pk %}">
+        <a class="button"
+           href="{% url 'organisaties:catalogi:producten:edit' pk=lokaleoverheid.pk catalog_pk=product.catalogus.pk product_pk=product.pk %}">
             {% trans "Wijzigen" %}
             {% if product.product_aanwezig %}
                 <i class="fa fa-sm fa-check text-accent" title="Dit product is beschikbaar."></i>
