+++ src/sdg/templates/producten/product_edit.html
@@ -134,7 +134,7 @@
                         </div>
                     </div>
                     <input class="button tabs__form-button" type="submit" value="Opslaan">
-                    <a href="{% url 'producten:detail' pk=product.pk %}">
+                    <a href="{% url 'organisaties:catalogi:producten:detail' pk=lokaleoverheid.pk catalog_pk=product.catalogus.pk product_pk=product.pk %}">
                         <button type="button" class="button button--light">{% trans 'Annuleren' %}</button>
                     </a>
                 </div>
